+++ examples/cache.go
@@ -8,6 +8,7 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
+// User describes a user
 type User struct {
 	Id   int64
 	Name string
@@ -38,7 +39,7 @@ func main() {
 		return
 	}
 
-	users := make([]User, 0)
+	var users []User
 	err = Orm.Find(&users)
 	if err != nil {
 		fmt.Println(err)
@@ -47,8 +48,7 @@ func main() {
 
 	fmt.Println("users:", users)
 
-	users2 := make([]User, 0)
-
+	var users2 []User
 	err = Orm.Find(&users2)
 	if err != nil {
 		fmt.Println(err)
@@ -57,8 +57,7 @@ func main() {
 
 	fmt.Println("users2:", users2)
 
-	users3 := make([]User, 0)
-
+	var users3 []User
 	err = Orm.Find(&users3)
 	if err != nil {
 		fmt.Println(err)
