+++ examples/conversion.go
@@ -9,35 +9,39 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
+// Status describes a status
 type Status struct {
 	Name  string
 	Color string
 }
 
+// defines some statuses
 var (
-	Registed = Status{"Registed", "white"}
-	Approved = Status{"Approved", "green"}
-	Removed  = Status{"Removed", "red"}
-	Statuses = map[string]Status{
-		Registed.Name: Registed,
-		Approved.Name: Approved,
-		Removed.Name:  Removed,
+	Registered = Status{"Registered", "white"}
+	Approved   = Status{"Approved", "green"}
+	Removed    = Status{"Removed", "red"}
+	Statuses   = map[string]Status{
+		Registered.Name: Registered,
+		Approved.Name:   Approved,
+		Removed.Name:    Removed,
 	}
 )
 
+// FromDB implemented xorm.Conversion convent database data to self
 func (s *Status) FromDB(bytes []byte) error {
 	if r, ok := Statuses[string(bytes)]; ok {
 		*s = r
 		return nil
-	} else {
-		return errors.New("no this data")
 	}
+	return errors.New("no this data")
 }
 
+// ToDB implemented xorm.Conversion convent to database data
 func (s *Status) ToDB() ([]byte, error) {
 	return []byte(s.Name), nil
 }
 
+// User describes a user
 type User struct {
 	Id     int64
 	Name   string
@@ -60,7 +64,7 @@ func main() {
 		return
 	}
 
-	_, err = Orm.Insert(&User{1, "xlw", Registed})
+	_, err = Orm.Insert(&User{1, "xlw", Registered})
 	if err != nil {
 		fmt.Println(err)
 		return
